+++ CHANGELOG.md
@@ -6,11 +6,11 @@ All important changes to this project will be documented in this file.
 
 #### Improvements:
 
-* --
+* Added fixing property `log-ips` in `server.properties` when secure authorization enabled on versions `1.20.2` and higher
 
 #### Fixed:
 
-* --
+* Fixed path comparison in process list filtering
 
 #### Breaking changes: